# haihang/swagger/spec/swagger.yaml
swagger: '2.0'
info:
  title: HNA支付模块接口
  description: 测试支付模块相关接口
  version: "1.2.0"
  contact: 
    email: bin.long@youruncloud.com
  termsOfService: http://gitlab.youruncloud.com/
# the domain of the service
# host: 192.168.3.222
host: somedomain
...
schemes:
  - http
# will be prefixed to all paths
basePath: /apis/v1
produces:
  - application/json
paths:
...
  /coupons/:
    post:
      summary: 创建优惠码
      description: |
        创建优惠码
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/coupon_create'
      tags:
        - coupons
      responses:
        201:
          description: 创建的优惠码信息
          schema:
            $ref: '#/definitions/coupon_create'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'

    get:
      summary: 优惠码列表
      description: |
        可用优惠码列表
      tags:
        - coupons
      responses:
        201:
          description: 优惠码列表信息
          schema:
            type: array
            items:
              $ref: '#/definitions/coupon'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /coupons/{coupon}/:
    get:
      summary: 获取指定优惠码详情
      description: |
        获取指定优惠码详情
      parameters:
        - name: coupon
          in: path
          description: 优惠码code或则uuid
          required: true
          type: string
      tags:
        - coupons
      responses:
        200:
          description: 指定优惠码对象
          schema:
            $ref: '#/definitions/coupon'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  coupon_create:
    type: object
    properties:
      create_count:
        type: number
        description: 创建数量
      coupon_type:
        type: string
        description: 优惠码类型, 有discount, reduce, recharge
      coupon_using_count:
        type: number
        description: 优惠码可使用次数
      coupon_value:
        type: number
        description: discount下<100, 其他模式为优惠码值(分)
      coupon_using_user:
        type: number
        description: 优惠码可供使用人数
      valid_datetime_start:
        type: string
        description: 优惠码有效开始时间, UTC
      valid_datetime_end:
        type: string
        description: 优惠码失效时间, UTC
  coupon:
    type: object
    properties:
      coupon_uuid:
        type: string
        description: 优惠码唯一标识
      coupon_code:
        type: string
        description: 优惠码code
      coupon_type:
        type: string
        description: 优惠码类型, 有discount, reduce, recharge
      coupon_using_count:
        type: number
        description: 优惠码可使用次数
      coupon_value:
        type: number
        description: discount下<100, 其他模式为优惠码值(分)
      coupon_using_user:
        type: number
        description: 优惠码可供使用人数
      using_user_left:
        type: number
        description: 优惠码剩余可使用人数
      valid_datetime_start:
        type: string
        description: 优惠码有效开始时间, UTC
      valid_datetime_end:
        type: string
        description: 优惠码失效时间, UTC
      create_time:
        type: string
        description: 优惠码创建时间
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string